% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{doHIP}
\alias{doHIP}
\title{Run the loop for Halton Iterative Partitioning}
\usage{
doHIP(dat2, n = NULL, quiet = TRUE, base = base, Jmax)
}
\arguments{
\item{dat2}{internal data table passed to the function.}

\item{n}{Sample size required}

\item{quiet}{whether or not to be print on each iteration.}

\item{base}{Co-prime bases 2,3}

\item{Maximum}{number of cuts by x,y J values.}
}
\description{
Run the loop for Halton Iterative Partitioning
}
