% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getBalance}
\alias{getBalance}
\title{Sample along a line using BAS.}
\usage{
getBalance(hip, n, smp = NULL)
}
\arguments{
\item{hip}{HIP exported data.table}

\item{n}{Number of samples}

\item{smp}{Spatial Points Sample, assumed sp.}
}
\description{
Sample along a line using BAS.
}
