% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getHBSamples}
\alias{getHBSamples}
\title{I honestly can't remember why I wrote this. I think it's an older version of getSamples. HB must stand for something afterall...}
\usage{
getHBSamples(dat, n, proj.string)
}
\arguments{
\item{dat}{HIP Exported data.table. Requires certain columns.}

\item{n}{sample size}

\item{proj.string}{Projection if needing to turn points into spatial points data frame for export.}
}
\description{
I honestly can't remember why I wrote this. I think it's an older version of getSamples. HB must stand for something afterall...
}
