% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getHIPBoxes}
\alias{getHIPBoxes}
\title{This function returns a bunch of polygons that represent the HIP splits of the population.
It really just does HIP again and could be combined with doHIP in the future with just the "boxes" 
added and create Polygon = TRUE or something. For now it's separate since it's plenty fast.}
\usage{
getHIPBoxes(hip, bb, n)
}
\arguments{
\item{hip}{HIP Exported data.table. Requires certain columns.}

\item{bb}{bounding box for this region.}

\item{n}{Number of HIP boxes to make.}
}
\description{
This function returns a bunch of polygons that represent the HIP splits of the population.
It really just does HIP again and could be combined with doHIP in the future with just the "boxes" 
added and create Polygon = TRUE or something. For now it's separate since it's plenty fast.
}
