% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getHipSample}
\alias{getHipSample}
\title{This is the Main Master Sample function that processes the data and then calls HIP and orders based the MSI and per stratum orderings.
Only  mess with this if you are writing your own master sample. Otherwise, we will have it in a NWT or NZ wrapper function for set
Master Sample orderings etc. It works on centroids of any discrete Halton space (rivers and lakes or just boxes from continuous space).}
\usage{
getHipSample(
  X,
  Y,
  index = NULL,
  N = NULL,
  bb,
  base = c(2, 3),
  seed = c(587, 8750),
  quiet = TRUE,
  Ps1 = 0:1,
  Ps2 = 0:2,
  hipS1 = 0:1,
  hipS2 = 0:2
)
}
\arguments{
\item{X}{Centroid coordinate for points be ordered x-axis.}

\item{Y}{Centroid coordinate for points be ordered y-axis.}

\item{index}{Master Sample Index for each point. If not supplied it will be calculated but slow things down and currently attemps to create 100 m boxes.}

\item{bb}{Master Sample bounding box.}

\item{base}{Generally 2,3. If you change it read the literature.}

\item{seed}{Master Sample Seed for two bases used to define the Master Sample Index.}

\item{Ps1}{Permutation for base 2 for calculating Master Sample Index.}

\item{Ps2}{Permutation for base 3 for calculating Master Sample Index.}

\item{hipS1}{Permutation of x orderings (0, 1) for this HIP (per stratum or individual sample)}

\item{hipS2}{Permutation of y orderings (0, 1, 2) for this HIP (per stratum or individual sample)}

\item{quite}{if you don't want to show each iteration of HIP keep this TRUE}
}
\description{
This is the Main Master Sample function that processes the data and then calls HIP and orders based the MSI and per stratum orderings.
Only  mess with this if you are writing your own master sample. Otherwise, we will have it in a NWT or NZ wrapper function for set
Master Sample orderings etc. It works on centroids of any discrete Halton space (rivers and lakes or just boxes from continuous space).
}
