% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getOverSamples}
\alias{getOverSamples}
\title{Written for Ollie to get over samples within a HIP box.}
\usage{
getOverSamples(dat, n, ni)
}
\arguments{
\item{dat}{HIP Exported data.table. Requires certain columns.}

\item{n}{sample size}

\item{ni}{Number of oversamples per HIP Box}
}
\description{
Written for Ollie to get over samples within a HIP box.
}
