% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getPolyRast}
\alias{getPolyRast}
\title{Create the Sample Raster:
This is for linear features. You pass the lines as a lines feature as well as the spatial sample smp. Then match by 
what was clipped in the Halton Boxes. Then draw a Halton sample along the lines. For continuous sampling inside a Halton box on a line.}
\usage{
getPolyRast(shp, smp, bb, J = c(5, 3), base = c(2, 3))
}
\arguments{
\item{shp}{Linear features for rivers or some sort of lines.}

\item{smp}{The chosen samples as data.table.}

\item{bb}{Master Sample bounding box.}

\item{J}{Master Sample cuts to figure out what the raster is.}

\item{base}{co-prime base. Almost always 2,3. Might really bugger things up if it weren't.}
}
\description{
Create the Sample Raster:
This is for linear features. You pass the lines as a lines feature as well as the spatial sample smp. Then match by 
what was clipped in the Halton Boxes. Then draw a Halton sample along the lines. For continuous sampling inside a Halton box on a line.
}
