% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getSamples}
\alias{getSamples}
\title{Select a sample given a HIP exported data.table.}
\usage{
getSamples(dat, n, lks = FALSE, proj.string = NULL)
}
\arguments{
\item{dat}{HIP Exported data.table. Requires certain columns.}

\item{n}{sample size}

\item{lks}{If lakes then change how spatial points are processed.}

\item{proj.string}{Projection if needing to turn points into spatial points data frame for export.}
}
\description{
Select a sample given a HIP exported data.table.
}
