% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{getVoronoi}
\alias{getVoronoi}
\title{This is a modified SDraw voronoi polygons code for the specific purposes of testing Spatial Balance with
objects typically built in this package.}
\usage{
getVoronoi(x, bbox)
}
\arguments{
\item{x}{is an sp spatial points}

\item{bbox}{the bounding box for the whole region.}
}
\description{
This is a modified SDraw voronoi polygons code for the specific purposes of testing Spatial Balance with
objects typically built in this package.
}
