% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWTMasterSample.R
\name{MasterSample}
\alias{MasterSample}
\title{Function to create Master Sample from points.}
\usage{
MasterSample(pts, B = 12, P1 = 0:1, P2 = 0:2, n = NULL)
}
\arguments{
\item{pts}{SpatialPointsDataFrame or SpatialPoints for pts that are to be sampled.}

\item{B}{Number of HIP Boxes to create for partitioning}

\item{P1}{Permutation for base 2 (0,1)}

\item{P2}{Permutation for base 3 (0,1,2)}

\item{n}{Sample size. Keep NULL if you want to return a shapefile with all points given the Master Sample Ordering.}
}
\description{
Function to create Master Sample from points.
}
