% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakes-data.r
\docType{data}
\name{NWT_Lakes}
\alias{NWT_Lakes}
\title{Data Processed as lake centroids for the GNWT CIMP Pilot Study}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(NWT_Lakes)
}
\description{
Included are all the lakes where a float plane can land > 1km,
}
\examples{
data(NWT_Lakes)
}
\keyword{datasets}
