% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{permutation_B}
\alias{permutation_B}
\title{Permute every level of Js but with the same random 0,1,2.}
\usage{
permutation_B(Ji = 3, bi = 2, si = c(1, 0))
}
\arguments{
\item{Ji}{integer for power of the base}

\item{bi}{the co-prime base}

\item{si}{random permutation of 0,1,2}
}
\description{
Permute every level of Js but with the same random 0,1,2.
}
