% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWTMasterSample.R
\name{plotMS}
\alias{plotMS}
\title{Plots a hip dataframe for sample Index into a leaflet map for easy visualization.}
\usage{
plotMS(pts, n)
}
\arguments{
\item{pts}{Spatial Points Data Fram exported from the master sample function.}

\item{n}{Number of points you want to plot.}
}
\description{
Plots a hip dataframe for sample Index into a leaflet map for easy visualization.
}
