% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{sampRiv}
\alias{sampRiv}
\title{Sample along a line using BAS.}
\usage{
sampRiv(n = 10, x, seed = 0)
}
\arguments{
\item{n}{Sample size}

\item{x}{Linear shape file clipped by a raster.}

\item{seed}{Random starting seed in one dimension for BAS linear features.}
}
\description{
Sample along a line using BAS.
}
