% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{subset_boxes}
\alias{subset_boxes}
\title{Do HIP but track the split points for making boxes.}
\usage{
subset_boxes(subdat, boxes, slice = 2)
}
\arguments{
\item{subdat}{HIP Exported data.table. Requires certain columns.}

\item{boxes}{passing the existing boxes to partition.}

\item{slice}{Which co-prime base are we partitioning meow.}
}
\description{
Do HIP but track the split points for making boxes.
}
