% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPMasterSample.r
\name{subset_dist}
\alias{subset_dist}
\title{Find the cut points for HIP loop.}
\usage{
subset_dist(n, xi, index, slice = 2)
}
\arguments{
\item{n}{an integer for number of points}

\item{xi}{is the coordinates}

\item{index}{Unique point ID}

\item{slice}{Which base to cut up.}
}
\description{
Find the cut points for HIP loop.
}
